Completed section8 + exercises
