+++ VidlyApp/Controllers/CustomersController.cs
@@ -10,7 +10,7 @@ using VidlyApp.ViewModels;
 
 namespace VidlyApp.Controllers
 {
- 
+    [Authorize]
     public class CustomersController : Controller
     {
