+++ VidlyApp/Models/ApplicationDbContext.cs
@@ -1,0 +1,24 @@
+﻿using System.Data.Entity;
+using System.Security.Claims;
+using System.Threading.Tasks;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.EntityFramework;
+using System.ComponentModel.DataAnnotations;
+using VidlyApp.Models;
+
+public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
+{
+    public DbSet<Customer> Customers { get; set; }
+    public DbSet<Movie> Movies { get; set; }
+    public DbSet<MembershipType> MembershipTypes { get; set; }
+    public DbSet<Gendre> Gendres { get; set; }
+    public ApplicationDbContext()
+        : base("DefaultConnection", throwIfV1Schema: false)
+    {
+    }
+
+    public static ApplicationDbContext Create()
+    {
+        return new ApplicationDbContext();
+    }
+}
