+++ VidlyApp/Models/ApplicationUser.cs
@@ -1,0 +1,25 @@
+﻿using System.Data.Entity;
+using System.Security.Claims;
+using System.Threading.Tasks;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.EntityFramework;
+using System.ComponentModel.DataAnnotations;
+
+// You can add profile data for the user by adding more properties to your ApplicationUser class, please visit http://go.microsoft.com/fwlink/?LinkID=317594 to learn more.
+public class ApplicationUser : IdentityUser
+{
+    [Required]
+    [StringLength(255)]
+    public string DrivingLicense { get; set; }
+
+    [Required]
+    public long Phone { get; set; }
+
+    public async Task<ClaimsIdentity> GenerateUserIdentityAsync(UserManager<ApplicationUser> manager)
+    {
+        // Note the authenticationType must match the one defined in CookieAuthenticationOptions.AuthenticationType
+        var userIdentity = await manager.CreateIdentityAsync(this, DefaultAuthenticationTypes.ApplicationCookie);
+        // Add custom user claims here
+        return userIdentity;
+    }
+}
