+++ VidlyApp/ViewModels/RegisterViewModel.cs
@@ -1,0 +1,28 @@
+﻿using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+
+public class RegisterViewModel
+{
+    [Required]
+    public long Phone { get; set; }
+
+    [Required]
+    [Display(Name = "Driving License")]
+    public string DrivingLicense { get; set; }
+
+    [Required]
+    [EmailAddress]
+    [Display(Name = "Email")]
+    public string Email { get; set; }
+
+    [Required]
+    [StringLength(100, ErrorMessage = "The {0} must be at least {2} characters long.", MinimumLength = 3)]
+    [DataType(DataType.Password)]
+    [Display(Name = "Password")]
+    public string Password { get; set; }
+
+    [DataType(DataType.Password)]
+    [Display(Name = "Confirm password")]
+    [Compare("Password", ErrorMessage = "The password and confirmation password do not match.")]
+    public string ConfirmPassword { get; set; }
+}
